created profiled generic and another profilled to oracle
